+++ documentation/app/src/index.js
@@ -3,10 +3,6 @@
 
 import { get } from "../c-at-e-file-server.js";
 import { env } from "fastly:env";
-import { KVStore } from "fastly:kv-store";
-
-// Pending update to https://github.com/JakeChampion/compute-file-server/commit/58bdcd4a6234bba5b74502f5f03bcaffabc7ad02
-globalThis.ObjectStore = KVStore;
 
 addEventListener("fetch", (event) => event.respondWith(app(event)));