+++ integration-tests/js-compute/fixtures/app/src/tee.js
@@ -1,5 +1,6 @@
 /* eslint-env serviceworker */
 import { routes } from './routes.js';
+import { assert } from './assertions.js';
 
 routes.set('/tee', async function (event) {
   const req = event.request;
@@ -13,21 +14,20 @@ routes.set('/tee', async function (event) {
   // `pending_req_select`, as we were waiting on an http request whose body had
   // not been closed.
   let res = await fetch(
-    new Request('/post', {
+    new Request('/anything', {
       body: body1,
       headers: req.headers,
       method: req.method,
-      backend: 'httpbin',
+      backend: 'httpme',
     }),
   );
   let body = await res.json();
-
-  return new Response(body.data);
+  assert(body['body'], 'hello world!');
 });
 
 routes.set('/tee/error', async function (event) {
   const req = event.request;
-  let res = fetch('/post', {
+  let res = fetch('/anything', {
     method: 'POST',
     body: new ReadableStream({
       start: (controller) => {
@@ -35,7 +35,7 @@ routes.set('/tee/error', async function (event) {
         controller.close();
       },
     }),
-    backend: 'httpbin',
+    backend: 'httpme',
   });
 
   return res