+++ runtime/fastly/builtins/fastly.h
@@ -60,10 +60,31 @@ public:
   static bool defaultBackend_set(JSContext *cx, unsigned argc, JS::Value *vp);
   static bool allowDynamicBackends_get(JSContext *cx, unsigned argc, JS::Value *vp);
   static bool allowDynamicBackends_set(JSContext *cx, unsigned argc, JS::Value *vp);
+  static bool inspect(JSContext *cx, unsigned argc, JS::Value *vp);
 };
 
 JS::Result<std::tuple<JS::UniqueChars, size_t>> convertBodyInit(JSContext *cx,
                                                                 JS::HandleValue bodyInit);
+
+inline bool get_fastly_object(api::Engine *engine, JS::MutableHandleObject out) {
+  JS::RootedValue fastly_val(engine->cx());
+  if (!JS_GetProperty(engine->cx(), engine->global(), "fastly", &fastly_val)) {
+    return false;
+  }
+  if (fastly_val.isObject()) {
+    out.set(&fastly_val.toObject());
+    return true;
+  }
+  JS::RootedObject fastly_obj(engine->cx(), JS_NewPlainObject(engine->cx()));
+  if (!fastly_obj) {
+    return false;
+  }
+  if (!JS_DefineProperty(engine->cx(), engine->global(), "fastly", fastly_obj, 0)) {
+    return false;
+  }
+  out.set(fastly_obj);
+  return true;
+}
 
 /**
  * Debug only logging system, adding messages to `fastly.debugMessages`