+++ types/index.d.ts
@@ -17,3 +17,7 @@
 /// <reference path="kv-store.d.ts" />
 /// <reference path="logger.d.ts" />
 /// <reference path="secret-store.d.ts" />
+/// <reference path="html-rewriter.d.ts" />
+/// <reference path="image-optimizer.d.ts" />
+/// <reference path="shielding.d.ts" />
+/// <reference path="security.d.ts" />